+++ src/rendering/vcInternalModels.cpp
@@ -1,0 +1,7 @@
+
+#include "vcInternalModels.h"
+
+const vcP1UV1Vertex *pTubeVertices = (vcP1UV1Vertex*)tubeVerticesFltArray;
+const vcP1UV1Vertex *pSphereVertices = (vcP1UV1Vertex*)sphereVerticesFltArray;
+const vcP1N1Vertex *pCompassVerts = (vcP1N1Vertex*)compassVertsFltArray;
+const vcP1N1Vertex *pOrbitVertices = (vcP1N1Vertex*)orbitVerticesFltArray;